abstract getParsedContractFunctionArgs func to uitlsContract and use it in ReadOnlyFunctionForm too
